% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexique}
\alias{lexique}
\title{Lexique}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 142694 rows and 35 columns.}
\usage{
lexique
}
\description{
Lexique
}
\keyword{datasets}
