% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montre_mots.R
\name{montre_mots}
\alias{montre_mots}
\title{Répéter les mots françaises plus communs}
\usage{
montre_mots(debout, fin = NULL, instructions = TRUE)
}
\arguments{
\item{debout}{int: Les mots plus hauts. Si vous utilisez `fin`, ceci le rang
inferieur.}

\item{fin}{int: Le rang superieur.}

\item{instructions}{Boolean: Montre les instructions ou non?}
}
\description{
Répéter les mots françaises plus communs
}
\examples{
# Les mille mots plus hauts
montre_mots(1000)

# Les mots de rang 500 à 600
montre_mots(500, 600)

# Les mille mots plus bas
montre_mots(-1000)
}
