% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mots}
\alias{mots}
\title{Mots}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 46947 rows and 3 columns.}
\usage{
mots
}
\description{
Mots
}
\keyword{datasets}
